% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sspower.3level.R
\name{sspower.3level}
\alias{sspower.3level}
\title{Sample Size and Power Calculations for Three-Level Cluster Randomized Trials with A Univariate Effect Modifier}
\usage{
sspower.3level(power = 0.8,
               nc = NULL,
               alpha = 0.05,
               pi = 0.5,
               sigma2x = 1,
               sigma2y = 1,
               rho0 = 0.02,
               rho1 = 0.02,
               alpha0 = 0.02,
               alpha1 = 0.02,
               m = 50,
               ns = 6,
               eff = 0.2,
               beta3 = 0.2,
               beta4 = 0.05,
               estimand = "HTE",
               random.level = "cluster",
               verbose = TRUE)
}
\arguments{
\item{power}{A numeric value between 0 and 1 specifying the desired power level for sample size estimation. Default is \code{0.8}.}

\item{nc}{A numeric value specifying the number of clusters provided by the user to estimate the achievable power. Default is \code{NULL}.}

\item{alpha}{A numeric value between 0 and 1 representing the type I error rate. Default is \code{0.05}.}

\item{pi}{A rational number between 0 and 1 indicating the proportion of treated units at the randomization level. Default is \code{0.5}, representing balanced allocation.}

\item{sigma2x}{A positive numeric value for the marginal variance of the univariate effect modifier. Default is \code{1}.}

\item{sigma2y}{A positive numeric value for the conditional variance of the continuous outcome. Default is \code{1}.}

\item{rho0}{A numeric value between 0 and 1 representing the intraclass correlation coefficient (ICC) for within-subcluster covariate variability. Default is \code{0.02}.}

\item{rho1}{A numeric value between 0 and 1 representing the ICC for between-subcluster covariate variability. Default is \code{0.02}.}

\item{alpha0}{A numeric value between 0 and 1 representing the ICC for within-subcluster outcome variability. Default is \code{0.02}.}

\item{alpha1}{A numeric value between 0 and 1 representing the ICC for between-subcluster outcome variability. Default is \code{0.02}.}

\item{m}{A numeric value greater than 2 specifying the number of participants per subcluster. Default is \code{50}.}

\item{ns}{A numeric value greater than 2 specifying the number of subclusters per cluster. Default is \code{6}.}

\item{eff}{A nonzero numeric value for the (unstandardized) effect size of the target estimand. The effect size can optionally be standardized relative to \code{sigma2y}, the
conditional variance of the outcome. Default is \code{0.1}.}

\item{beta3}{A nonzero numeric value specifying the true covariate effect, corresponding to the model coefficient of the covariate.
Default is \code{0.2}. This parameter is required only when the estimand is \code{"unATE"}.}

\item{beta4}{A nonzero numeric value specifying the true effect of treatment effect heterogeneity, corresponding to the model coefficient
of the covariate-treatment interaction. Default is \code{0.05}. This parameter is required only when the estimand is \code{"unATE"}.}

\item{estimand}{A character string indicating the type of treatment effect estimand. Supported values are \code{"HTE"} (heterogeneous treatment effect),
\code{"ATE"} (average treatment effect), and \code{"unATE"} (unadjusted average treatment effect). Default is \code{"HTE"}.}

\item{random.level}{A character string specifying the level of randomization. Supported values are \code{"cluster"} (randomization at the cluster level),
\code{"subcluster"} (randomization at the subcluster level), and \code{"individual"} (individual-level randomization). Default is \code{"cluster"}.}

\item{verbose}{A logical value indicating whether parameter reiterations and supplementary messages should be displayed (\code{TRUE}) or suppressed (\code{FALSE}). Default is \code{TRUE}.}
}
\value{
The function \code{sspower.multi.3level} returns one of two outputs: either an integer specifying the required number of clusters along with the actual power achieved
(slightly larger than the specified power), or a decimal value representing the achieved power based on the provided sample size. The power is calculated to four decimal places.
Informative messages summarizing key parameter choices and results are displayed by default.
}
\description{
The function \code{sspower.3level} calculates either the required number of clusters or the achieved power for three-level cluster randomized
trials (CRTs) with a univariate effect modifier. It accommodates randomization at the cluster, subcluster, or individual level and supports hypothesis testing
for three types of effects: the heterogeneous treatment effect (HTE), the average treatment effect (ATE), and the unadjusted ATE (estimated via a linear mixed model
without adjusting for the effect modifier). In particular, the sample size calculations and power analysis for testing the ATE or unadjusted ATE are independent of
the effect modifier, meaning that the dimension of the effect modifiers does not affect the results. Users can specify the number of clusters (level-3) using the
\code{nc} argument. If this argument is provided, the function calculates power for the specified hypothesis test and disregards any input for the power parameter.
If the number of clusters is not provided, the function computes the required number of clusters to achieve a specified power threshold, which defaults to 0.8.
}
\details{
Based on the input parameters, the method first calculates the variances of the effects of interest using Generalized Least Squares estimators and large-sample approximations.
These variances are then applied to construct either classic sample size formulas or power formulas, facilitating both sample size and power calculations. Specifically, normal quantiles
are employed for sample size calculations, except in the case of cluster-level randomization with ATE and unATE estimands, where t-quantiles are used to incorporate small-sample adjustments.
The methods support three estimands (HTE, ATE, and unATE) and three randomization levels (cluster, subcluster, and individual), as detailed in the paper entitled "Planning Three-Level Cluster
Randomized Trials to Assess Treatment Effect Heterogeneity" by Li et al.
}
\examples{
# Predict the achieved power for 20 clusters with default parameter values.
power.example <- sspower.3level(power = 0.8,
                                nc = 20,
                                alpha = 0.05,
                                pi = 0.5,
                                sigma2x = 1,
                                sigma2y = 1,
                                rho0 = 0.02,
                                rho1 = 0.02,
                                alpha0 = 0.02,
                                alpha1 = 0.02,
                                m = 50,
                                ns = 6,
                                eff = 0.2,
                                beta3 = 0.2,
                                beta4 = 0.05,
                                estimand = "HTE",
                                random.level = "cluster",
                                verbose = TRUE)

print(power.example)

}
