% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sspower.multi.SW.R
\name{sspower.multi.SW}
\alias{sspower.multi.SW}
\title{Sample Size and Power Calculations for Stepped Wedge Cluster Randomized Trials with A Multivariate Effect Modifier}
\usage{
sspower.multi.SW(power = 0.8,
                 I = NULL,
                 alpha = 0.05,
                 sigma2x = diag(rep(1,3)),
                 sigma2y = 1,
                 rho0 = diag(rep(0.02,3)),
                 rho1 = diag(rep(0.02,3)),
                 alpha0 = 0.02,
                 alpha1 = 0.02,
                 alpha2 = 0.02,
                 N = 50,
                 J = 6,
                 eff = rep(0.5,3),
                 design = "cross-sectional",
                 verbose = TRUE)
}
\arguments{
\item{power}{A numeric value between 0 and 1 specifying the desired power level for sample size estimation. Default is \code{0.8}.}

\item{I}{A numeric value specifying the number of clusters provided by the user to estimate the achievable power. Default is \code{NULL}.}

\item{alpha}{A numeric value between 0 and 1 representing the type I error rate. Default is \code{0.05}.}

\item{sigma2x}{A positive-definite covariance matrix for the marginal effect modifiers. Default is \code{diag(rep(1,3))}.}

\item{sigma2y}{A positive numeric value for the conditional variance of the continuous outcome. Default is \code{1}.}

\item{rho0}{A cross-correlation matrix with elements between 0 and 1 representing the intraclass correlation coefficients (ICCs) for within-period covariates. Default is \code{diag(0.02,3)}.}

\item{rho1}{A cross-correlation matrix with elements between 0 and 1 representing the intraclass correlation coefficients (ICCs) for between-period covariates. Default is \code{diag(0.02,3)}.}

\item{alpha0}{A numeric value between 0 and 1 representing the ICC for within-period outcome variability. Default is \code{0.02}.}

\item{alpha1}{A numeric value between 0 and 1 representing the ICC for between-period outcome variability. Default is \code{0.02}.}

\item{alpha2}{A numeric value between 0 and 1 representing the ICC for within-individual outcome variability, which is only required for closed-cohort design. Default is \code{0.02}.}

\item{N}{A numeric value greater than 2 specifying the number of individuals per period. Default is \code{50}.}

\item{J}{A numeric value greater than 2 specifying the number of periods per cluster. Default is \code{6}.}

\item{eff}{A nonzero vector of numeric values for the (unstandardized) effect size of the target estimand. The effect size can optionally be standardized relative to \code{sigma2y}, the
conditional variance of the outcome. Default is \code{rep(0.5,3)}.}

\item{design}{A character string indicating the type of stepped wedge design. Supported values are \code{"cross-sectional"} (cross-sectional design) and
\code{"closed-cohort"} (closed-cohort design). Default is \code{"cross-sectional"}.}

\item{verbose}{A logical value indicating whether parameter reiterations and supplementary messages should be displayed (\code{TRUE}) or suppressed (\code{FALSE}). Default is \code{TRUE}.}
}
\value{
The function \code{sspower.multi.SW} returns one of two outputs: either an integer specifying the required number of clusters along with the actual power achieved
(slightly larger than the specified power), or a decimal value representing the achieved power based on the provided sample size. The power is calculated to four decimal places.
Informative messages summarizing key parameter choices and results are displayed by default.
}
\description{
The \code{sspower.multi.SW} function calculates sample size or power for stepped wedge cluster randomized trials (SW-CRTs) with a multivariate effect modifier.
It supports both cross-sectional and closed-cohort stepped wedge designs and focuses on hypothesis testing for heterogeneous treatment effects (HTE).
Users can specify the number of clusters using the \code{I} argument. When \code{I} is provided, the function computes the power for the specified hypothesis
test, disregarding the \code{power} parameter. If \code{I} is not specified, the function calculates the required number of clusters to achieve the desired power
threshold, which defaults to 0.8.
}
\details{
This method employs Generalized Least Squares estimators and large-sample approximations to compute the variances of the effects of interest.
These variances are then utilized to derive formulas for sample size and power calculations, supporting analysis for both objectives.
The approach accommodates two stepped wedge designs—cross-sectional and closed-cohort—with a primary focus on heterogeneous treatment effects (HTE).
As described in the paper "Planning Stepped Wedge Cluster Randomized Trials to Detect Treatment Effect Heterogeneity" by Li et al.,
the sample size and power calculations align with those produced by \code{sspower.SW} when the target estimand is the average treatment effect,
as the intraclass correlation coefficients (ICCs) of the covariates do not affect these computations.
}
\examples{
# Predict the achieved power for 20 clusters with default parameter values.
power.example <- sspower.multi.SW(power = 0.8,
                                  I = 20,
                                  alpha = 0.05,
                                  sigma2x = diag(rep(1,3)),
                                  sigma2y = 1,
                                  rho0 = diag(0.02,3),
                                  rho1 = diag(0.02,3),
                                  alpha0 = 0.02,
                                  alpha1 = 0.02,
                                  alpha2 = 0.02,
                                  N = 50,
                                  J = 6,
                                  eff = rep(0.5,3),
                                  design = "cross-sectional",
                                  verbose = TRUE)

print(power.example)

}
