% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sspower.SW.R
\name{sspower.SW}
\alias{sspower.SW}
\title{Sample Size and Power Calculations for Stepped Wedge Cluster Randomized Trials with A Univariate Effect Modifier}
\usage{
sspower.SW(power = 0.8,
           I = NULL,
           alpha = 0.05,
           sigma2x = 1,
           sigma2y = 1,
           rho0 = 0.02,
           rho1 = 0.02,
           alpha0 = 0.02,
           alpha1 = 0.02,
           alpha2 = 0.02,
           N = 50,
           J = 6,
           eff = 0.1,
           estimand = "HTE",
           design = "cross-sectional",
           verbose = TRUE)
}
\arguments{
\item{power}{A numeric value between 0 and 1 specifying the desired power level for sample size estimation. Default is \code{0.8}.}

\item{I}{A numeric value specifying the number of clusters provided by the user to estimate the achievable power. Default is \code{NULL}.}

\item{alpha}{A numeric value between 0 and 1 representing the type I error rate. Default is \code{0.05}.}

\item{sigma2x}{A positive numeric value for the marginal variance of the univariate effect modifier. Default is \code{1}.}

\item{sigma2y}{A positive numeric value for the conditional variance of the continuous outcome. Default is \code{1}.}

\item{rho0}{A numeric value between 0 and 1 representing the intraclass correlation coefficient (ICC) for within-period covariate variability. Default is \code{0.02}.}

\item{rho1}{A numeric value between 0 and 1 representing the ICC for between-period covariate variability. Default is \code{0.02}.}

\item{alpha0}{A numeric value between 0 and 1 representing the ICC for within-period outcome variability. Default is \code{0.02}.}

\item{alpha1}{A numeric value between 0 and 1 representing the ICC for between-period outcome variability. Default is \code{0.02}.}

\item{alpha2}{A numeric value between 0 and 1 representing the ICC for within-individual outcome variability, which is only required for closed-cohort design. Default is \code{0.02}.}

\item{N}{A numeric value greater than 2 specifying the number of individuals per period. Default is \code{50}.}

\item{J}{A numeric value greater than 2 specifying the number of periods per cluster. Default is \code{6}.}

\item{eff}{A nonzero numeric value for the (unstandardized) effect size of the target estimand. The effect size can optionally be standardized relative to \code{sigma2y}, the
conditional variance of the outcome. Default is \code{0.1}.}

\item{estimand}{A character string indicating the type of treatment effect estimand. Supported values are \code{"HTE"} (heterogeneous treatment effect) and
\code{"ATE"} (average treatment effect). Default is \code{"HTE"}.}

\item{design}{A character string indicating the type of stepped wedge design. Supported values are \code{"cross-sectional"} (cross-sectional design) and
\code{"closed-cohort"} (closed-cohort design). Default is \code{"cross-sectional"}.}

\item{verbose}{A logical value indicating whether parameter reiterations and supplementary messages should be displayed (\code{TRUE}) or suppressed (\code{FALSE}). Default is \code{TRUE}.}
}
\value{
The function \code{sspower.SW} returns one of two outputs: either an integer specifying the required number of clusters along with the actual power achieved
(slightly larger than the specified power), or a decimal value representing the achieved power based on the provided sample size. The power is calculated to four decimal places.
Informative messages summarizing key parameter choices and results are displayed by default.
}
\description{
The \code{sspower.SW} function calculates sample size or power for stepped wedge cluster randomized trials (SW-CRTs) with a univariate effect modifier.
It supports both cross-sectional and closed-cohort stepped wedge designs and allows hypothesis testing for heterogeneous treatment effects (HTE)
and average treatment effects (ATE). Users can specify the number of clusters using the \code{I} argument. When \code{I} is
provided, the function computes the power for the specified hypothesis test, disregarding the \code{power} parameter. If \code{I} is not specified,
the function calculates the required number of clusters to achieve the desired power threshold, which defaults to 0.8.
}
\details{
The method leverages Generalized Least Squares estimators and large-sample approximations to compute the variances of the effects of interest.
These variances are subsequently applied to derive sample size and power formulas, enabling calculations for both objectives.
The approach supports two estimands (HTE and ATE) and two designs (cross-sectional and closed-cohort),
as outlined in the paper "Planning Stepped Wedge Cluster Randomized Trials to Detect Treatment Effect Heterogeneity" by Li et al.
For sample size calculations, normal quantiles are used when the target estimand is the HTE, while t-quantiles are employed for small-sample
adjustments when the target estimand is the ATE.
}
\examples{
# Predict the achieved power for 20 clusters with default parameter values.
power.example <- sspower.SW(power = 0.8,
                            I = 20,
                            alpha = 0.05,
                            sigma2x = 1,
                            sigma2y = 1,
                            rho0 = 0.02,
                            rho1 = 0.02,
                            alpha0 = 0.02,
                            alpha1 = 0.02,
                            alpha2 = 0.02,
                            N = 50,
                            J = 6,
                            eff = 0.1,
                            estimand = "HTE",
                            design = "cross-sectional",
                            verbose = TRUE)

print(power.example)

}
